backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.10.0
google-generativeai==0.8.3
pydantic==2.10.5
requests==2.31.0
